continued work on linux porting
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=b6cbcd1c65ad9fdba24436d0b4c5ecd76a7b7694

+++ code/components/net-base/src/NetUdpSocket.cpp
@@ -35,7 +35,7 @@ bool UdpSocket::Bind(const PeerAddress& localAddress)
 {
 	if (localAddress.GetAddressFamily() != (int)m_addressFamily)
 	{
-		trace("Failed to bind socket - address family %d does not match %d.\n", m_addressFamily, localAddress.GetAddressFamily());
+		trace("Failed to bind socket - address family %d does not match %d.\n", (int)m_addressFamily, localAddress.GetAddressFamily());
 		return false;
 	}
